+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # AssetMate Changelog
 
-## Unreleased
+## 2.4.0 - 2024-01-15
 ### Added
-- Added the `assetsmate/purge` CLI command for bulk-deleting unused assets
-- Added the `assetsmate/purge/folders` CLI command for bulk-deleting empty folders
+- Added the `assetmate/purge` CLI command for bulk-deleting unused assets
+- Added the `assetmate/purge/folders` CLI command for bulk-deleting empty folders
 
 ## 2.3.2 - 2023-09-28
 ### Fixed